test(retrofit): moved http tests to new SpinnakerHttpExceptionTest class
